+++ src/CMakeLists.txt
@@ -1,2 +1,2 @@
-target_sources(app PRIVATE main.c hog.c revsvc.c batterylvl.c button.c)
+target_sources(app PRIVATE main.c hog.c revsvc.c batterylvl.c button.c ble.c)
 target_include_directories(app PRIVATE .)
