+++ src/button.c
@@ -1,5 +1,6 @@
 
 #include "button.h"
+#include "ble.h"
 // thread stuff
 
 k_tid_t button_thread_tid;
@@ -39,7 +40,8 @@ void button_thread_fn(void *arg1, void *arg2, void *arg3) {
 
             // Debounce delay
             k_sleep(K_MSEC(500));
-            // k_work_submit(&advertise_acceptlist_work);
+            bluetooth_adv();
+
         }
 
         k_sleep(K_MSEC(100));  // Polling delay
